requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
langchain
langchain-core
langchain-openai
openai
python-dotenv
PyGithub